[compiler] move TableIR execute methods out of class
