+++ hail/hail/src/is/hail/expr/ir/TableIR.scala
@@ -4,8 +4,7 @@ import is.hail.HailContext
 import is.hail.annotations._
 import is.hail.asm4s._
 import is.hail.backend.{ExecuteContext, HailStateManager, HailTaskContext, TaskFinalizer}
-import is.hail.backend.spark.{SparkBackend, SparkTaskContext}
-import is.hail.expr.ir.compile.{Compile, CompileWithAggregators}
+import is.hail.expr.ir.compile.Compile
 import is.hail.expr.ir.defs._
 import is.hail.expr.ir.functions.{
   BlockMatrixToTableFunction, IntervalFunctions, MatrixToTableFunction, TableToTableFunction,
@@ -27,13 +26,9 @@ import is.hail.types.physical.stypes.interfaces._
 import is.hail.types.physical.stypes.primitives.{SInt64, SInt64Value}
 import is.hail.types.virtual._
 import is.hail.utils._
-import is.hail.utils.prettyPrint.ArrayOfByteArrayInputStream
 
-import scala.reflect.ClassTag
+import java.io.{Closeable, InputStream}
 
-import java.io.{Closeable, DataInputStream, DataOutputStream, InputStream}
-
-import org.apache.spark.TaskContext
 import org.apache.spark.sql.Row
 import org.json4s.{DefaultFormats, Extraction, Formats, JValue, ShortTypeHints}
 import org.json4s.JsonAST.JString
@@ -61,15 +56,6 @@ sealed abstract class TableIR extends BaseIR {
   def partitionCounts: Option[IndexedSeq[Long]] = None
 
   val rowCountUpperBound: Option[Long]
-
-  final def analyzeAndExecute(ctx: ExecuteContext): TableExecuteIntermediate = {
-    val r = Requiredness(this, ctx)
-    val d = DistinctlyKeyed(this)
-    execute(ctx, LoweringAnalyses(r, d))
-  }
-
-  protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses): TableExecuteIntermediate =
-    fatal("tried to execute unexecutable IR:\n" + Pretty(ctx, this))
 
   override protected def copyWithNewChildren(newChildren: IndexedSeq[BaseIR]): TableIR
 
@@ -108,20 +94,6 @@ case class TableLiteral(
   override protected def copyWithNewChildren(newChildren: IndexedSeq[BaseIR]): TableLiteral = {
     assert(newChildren.isEmpty)
     TableLiteral(typ, rvd, enc, encodedGlobals)
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val (globalPType: PStruct, dec) = enc.buildDecoder(ctx, typ.globalType)
-
-    val bais = new ArrayOfByteArrayInputStream(encodedGlobals)
-    val globalOffset = dec.apply(bais, ctx.theHailClassLoader).readRegionValue(ctx.r)
-    new TableValueIntermediate(TableValue(
-      ctx,
-      typ,
-      BroadcastRow(ctx, RegionValue(ctx.r, globalOffset), globalPType),
-      rvd,
-    ))
   }
 }
 
@@ -2132,10 +2104,6 @@ case class TableRead(typ: TableType, dropRows: Boolean, tr: TableReader) extends
     assert(newChildren.isEmpty)
     TableRead(typ, dropRows, tr)
   }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate =
-    tr.toExecuteIntermediate(ctx, typ, dropRows)
 }
 
 case class TableParallelize(rowsAndGlobal: IR, nPartitions: Option[Int] = None) extends TableIR {
@@ -2163,65 +2131,6 @@ case class TableParallelize(rowsAndGlobal: IR, nPartitions: Option[Int] = None) 
       FastSeq(),
       globalsType,
     )
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val (ptype: PStruct, res) =
-      CompileAndEvaluate._apply(ctx, rowsAndGlobal, optimize = false) match {
-        case Right((t, off)) => (t.fields(0).typ, t.loadField(off, 0))
-      }
-
-    val globalsT = ptype.types(1).setRequired(true).asInstanceOf[PStruct]
-    if (ptype.isFieldMissing(res, 1))
-      fatal("'parallelize': found missing global value")
-    val globals = BroadcastRow(ctx, RegionValue(ctx.r, ptype.loadField(res, 1)), globalsT)
-
-    val rowsT = ptype.types(0).asInstanceOf[PArray]
-    val rowT = rowsT.elementType.asInstanceOf[PStruct].setRequired(true)
-    val spec = TypedCodecSpec(ctx, rowT, BufferSpec.wireSpec)
-
-    val makeEnc = spec.buildEncoder(ctx, rowT)
-    val rowsAddr = ptype.loadField(res, 0)
-    val nRows = rowsT.loadLength(rowsAddr)
-
-    val nSplits = math.min(nPartitions.getOrElse(16), math.max(nRows, 1))
-    val parts = partition(nRows, nSplits)
-
-    val bae = new ByteArrayEncoder(ctx.theHailClassLoader, makeEnc)
-    var idx = 0
-    val encRows = Array.tabulate(nSplits) { splitIdx =>
-      val n = parts(splitIdx)
-      bae.reset()
-      val stop = idx + n
-      while (idx < stop) {
-        if (rowsT.isElementMissing(rowsAddr, idx))
-          fatal(s"cannot parallelize null values: found null value at index $idx")
-        bae.writeRegionValue(ctx.r, rowsT.loadElement(rowsAddr, idx))
-        idx += 1
-      }
-      (n, bae.result())
-    }
-
-    val (resultRowType: PStruct, makeDec) = spec.buildDecoder(ctx, typ.rowType)
-    assert(
-      resultRowType.virtualType == typ.rowType,
-      s"typ mismatch:" +
-        s"\n  res=${resultRowType.virtualType}\n  typ=${typ.rowType}",
-    )
-
-    log.info(s"parallelized $nRows rows in $nSplits partitions")
-
-    val rvd = ContextRDD.parallelize(encRows, encRows.length)
-      .cmapPartitions { (ctx, it) =>
-        it.flatMap { case (nRowPartition, arr) =>
-          val bais = new ByteArrayDecoder(theHailClassLoaderForSparkWorkers, makeDec)
-          bais.set(arr)
-          Iterator.range(0, nRowPartition)
-            .map(_ => bais.readValue(ctx.region))
-        }
-      }
-    new TableValueIntermediate(TableValue(ctx, typ, globals, RVD.unkeyed(resultRowType, rvd)))
   }
 }
 
@@ -2254,12 +2163,6 @@ case class TableKeyBy(child: TableIR, keys: IndexedSeq[String], isSorted: Boolea
   override protected def copyWithNewChildren(newChildren: IndexedSeq[BaseIR]): TableKeyBy = {
     assert(newChildren.length == 1)
     TableKeyBy(newChildren(0).asInstanceOf[TableIR], keys, isSorted)
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val tv = child.execute(ctx, r).asTableValue(ctx)
-    TableValueIntermediate(tv.copy(typ = typ, rvd = tv.rvd.enforceKey(ctx, keys, isSorted)))
   }
 }
 
@@ -2325,15 +2228,6 @@ case class TableGen(
 
   override def childrenSeq: IndexedSeq[BaseIR] =
     FastSeq(contexts, globals, body)
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate =
-    new TableStageIntermediate(LowerTableIR.applyTable(
-      this,
-      DArrayLowering.All,
-      ctx,
-      LoweringAnalyses(this, ctx),
-    ))
 }
 
 case class TableRange(n: Int, nPartitions: Int) extends TableIR {
@@ -2349,7 +2243,8 @@ case class TableRange(n: Int, nPartitions: Int) extends TableIR {
 
   private val partCounts = partition(n, nPartitionsAdj)
 
-  override val partitionCounts = Some(partCounts.map(_.toLong).toFastSeq)
+  override val partitionCounts: Some[IndexedSeq[Long]] =
+    Some(partCounts.map(_.toLong).toFastSeq)
 
   lazy val rowCountUpperBound: Option[Long] = Some(n.toLong)
 
@@ -2358,44 +2253,6 @@ case class TableRange(n: Int, nPartitions: Int) extends TableIR {
     Array("idx"),
     TStruct.empty,
   )
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val localRowType = PCanonicalStruct(true, "idx" -> PInt32Required)
-    val localPartCounts = partCounts
-    val partStarts = partCounts.scanLeft(0)(_ + _)
-    new TableValueIntermediate(TableValue(
-      ctx,
-      typ,
-      BroadcastRow.empty(ctx),
-      new RVD(
-        RVDType(localRowType, Array("idx")),
-        new RVDPartitioner(
-          ctx.stateManager,
-          Array("idx"),
-          typ.rowType,
-          Array.tabulate(nPartitionsAdj) { i =>
-            val start = partStarts(i)
-            val end = partStarts(i + 1)
-            Interval(Row(start), Row(end), includesStart = true, includesEnd = false)
-          },
-        ),
-        ContextRDD.parallelize(Range(0, nPartitionsAdj), nPartitionsAdj)
-          .cmapPartitionsWithIndex { case (i, ctx, _) =>
-            val region = ctx.region
-
-            val start = partStarts(i)
-            Iterator.range(start, start + localPartCounts(i))
-              .map { j =>
-                val off = localRowType.allocate(region)
-                localRowType.setFieldPresent(off, 0)
-                Region.storeInt(localRowType.fieldOffset(off, 0), j)
-                off
-              }
-          },
-      ),
-    ))
-  }
 }
 
 case class TableFilter(child: TableIR, pred: IR) extends TableIR {
@@ -2408,39 +2265,6 @@ case class TableFilter(child: TableIR, pred: IR) extends TableIR {
   override protected def copyWithNewChildren(newChildren: IndexedSeq[BaseIR]): TableFilter = {
     assert(newChildren.length == 2)
     TableFilter(newChildren(0).asInstanceOf[TableIR], newChildren(1).asInstanceOf[IR])
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val tv = child.execute(ctx, r).asTableValue(ctx)
-
-    if (pred == True())
-      return TableValueIntermediate(tv)
-    else if (pred == False())
-      return TableValueIntermediate(tv.copy(rvd = RVD.empty(ctx, typ.canonicalRVDType)))
-
-    val (Some(BooleanSingleCodeType), f) = Compile[AsmFunction3RegionLongLongBoolean](
-      ctx,
-      FastSeq(
-        (
-          TableIR.rowName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.rvd.rowPType)),
-        ),
-        (
-          TableIR.globalName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.globals.t)),
-        ),
-      ),
-      FastSeq(classInfo[Region], LongInfo, LongInfo),
-      BooleanInfo,
-      Coalesce(FastSeq(pred, False())),
-    )
-
-    TableValueIntermediate(
-      tv.filterWithPartitionOp(ctx.theHailClassLoader, ctx.fsBc, f)((rowF, ctx, ptr, globalPtr) =>
-        rowF(ctx.region, ptr, globalPtr)
-      )
-    )
   }
 }
 
@@ -2467,15 +2291,6 @@ trait TableSubset extends TableIR {
   lazy val rowCountUpperBound: Option[Long] = child.rowCountUpperBound match {
     case Some(c) => Some(c.min(n))
     case None => Some(n)
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val prev = child.execute(ctx, r).asTableValue(ctx)
-    new TableValueIntermediate(prev.copy(rvd = subsetKind match {
-      case TableSubset.HEAD => prev.rvd.head(n, child.partitionCounts)
-      case TableSubset.TAIL => prev.rvd.tail(n, child.partitionCounts)
-    }))
   }
 }
 
@@ -2515,18 +2330,6 @@ case class TableRepartition(child: TableIR, n: Int, strategy: Int) extends Table
   override protected def copyWithNewChildren(newChildren: IndexedSeq[BaseIR]): TableRepartition = {
     val IndexedSeq(newChild: TableIR) = newChildren
     TableRepartition(newChild, n, strategy)
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val prev = child.execute(ctx, r).asTableValue(ctx)
-    val rvd = strategy match {
-      case RepartitionStrategy.SHUFFLE => prev.rvd.coalesce(ctx, n, shuffle = true)
-      case RepartitionStrategy.COALESCE => prev.rvd.coalesce(ctx, n, shuffle = false)
-      case RepartitionStrategy.NAIVE_COALESCE => prev.rvd.naiveCoalesce(n, ctx)
-    }
-
-    new TableValueIntermediate(prev.copy(rvd = rvd))
   }
 }
 
@@ -2605,13 +2408,6 @@ case class TableJoin(left: TableIR, right: TableIR, joinType: String, joinKey: I
       joinKey,
     )
   }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val leftTV = left.execute(ctx, r).asTableStage(ctx)
-    val rightTV = right.execute(ctx, r).asTableStage(ctx)
-    TableExecuteIntermediate(LowerTableIRHelpers.lowerTableJoin(ctx, r, this, leftTV, rightTV))
-  }
 }
 
 case class TableIntervalJoin(
@@ -2638,86 +2434,6 @@ case class TableIntervalJoin(
     )
 
   override def partitionCounts: Option[IndexedSeq[Long]] = left.partitionCounts
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val leftValue = left.execute(ctx, r).asTableValue(ctx)
-    val rightValue = right.execute(ctx, r).asTableValue(ctx)
-
-    val leftRVDType = leftValue.rvd.typ
-    val rightRVDType = rightValue.rvd.typ.copy(key = rightValue.typ.key)
-    val rightValueFields = rightRVDType.valueType.fieldNames
-
-    val sm = ctx.stateManager
-    val localKey = typ.key
-    val localRoot = root
-    val newRVD =
-      if (product) {
-        val joiner = (rightPType: PStruct) => {
-          val leftRowType = leftRVDType.rowType
-          val newRowType = leftRowType.appendKey(
-            localRoot,
-            PCanonicalArray(rightPType.selectFields(rightValueFields)),
-          )
-          (
-            RVDType(newRowType, localKey),
-            (_: RVDContext, it: Iterator[Muple[RegionValue, Iterable[RegionValue]]]) => {
-              val rvb = new RegionValueBuilder(sm)
-              val rv2 = RegionValue()
-              it.map { case Muple(rv, is) =>
-                rvb.set(rv.region)
-                rvb.start(newRowType)
-                rvb.startStruct()
-                rvb.addAllFields(leftRowType, rv)
-                rvb.startArray(is.size)
-                is.foreach(i => rvb.selectRegionValue(rightPType, rightRVDType.valueFieldIdx, i))
-                rvb.endArray()
-                rvb.endStruct()
-                rv2.set(rv.region, rvb.end())
-
-                rv2
-              }
-            },
-          )
-        }
-
-        leftValue.rvd.orderedLeftIntervalJoin(ctx, rightValue.rvd, joiner)
-      } else {
-        val joiner = (rightPType: PStruct) => {
-          val leftRowType = leftRVDType.rowType
-          val newRowType = leftRowType.appendKey(
-            localRoot,
-            rightPType.selectFields(rightValueFields).setRequired(false),
-          )
-
-          (
-            RVDType(newRowType, localKey),
-            (_: RVDContext, it: Iterator[JoinedRegionValue]) => {
-              val rvb = new RegionValueBuilder(sm)
-              val rv2 = RegionValue()
-              it.map { case Muple(rv, i) =>
-                rvb.set(rv.region)
-                rvb.start(newRowType)
-                rvb.startStruct()
-                rvb.addAllFields(leftRowType, rv)
-                if (i == null)
-                  rvb.setMissing()
-                else
-                  rvb.selectRegionValue(rightPType, rightRVDType.valueFieldIdx, i)
-                rvb.endStruct()
-                rv2.set(rv.region, rvb.end())
-
-                rv2
-              }
-            },
-          )
-        }
-
-        leftValue.rvd.orderedLeftIntervalJoinDistinct(ctx, rightValue.rvd, joiner)
-      }
-
-    new TableValueIntermediate(TableValue(ctx, typ, leftValue.globals, newRVD))
-  }
 }
 
 /** The TableMultiWayZipJoin node assumes that input tables have distinct keys. If inputs do not
@@ -2745,102 +2461,19 @@ case class TableMultiWayZipJoin(
 
   lazy val rowCountUpperBound: Option[Long] = None
 
-  private def newGlobalType = TStruct(globalName -> TArray(first.typ.globalType))
-  private def newValueType = TStruct(fieldName -> TArray(first.typ.valueType))
-  private def newRowType = first.typ.keyType ++ newValueType
-
-  lazy val typ: TableType = first.typ.copy(
-    rowType = newRowType,
-    globalType = newGlobalType,
-  )
+  lazy val typ: TableType = {
+    def newGlobalType = TStruct(globalName -> TArray(first.typ.globalType))
+    def newValueType = TStruct(fieldName -> TArray(first.typ.valueType))
+    def newRowType = first.typ.keyType ++ newValueType
+    first.typ.copy(
+      rowType = newRowType,
+      globalType = newGlobalType,
+    )
+  }
 
   override protected def copyWithNewChildren(newChildren: IndexedSeq[BaseIR])
     : TableMultiWayZipJoin =
     TableMultiWayZipJoin(newChildren.asInstanceOf[IndexedSeq[TableIR]], fieldName, globalName)
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val sm = ctx.stateManager
-    val childValues = childrenSeq.map(_.execute(ctx, r).asTableValue(ctx))
-
-    val childRVDs = RVD.unify(ctx, childValues.map(_.rvd)).toFastSeq
-    assert(childRVDs.forall(_.typ.key.startsWith(typ.key)))
-
-    val repartitionedRVDs =
-      if (
-        childRVDs(0).partitioner.satisfiesAllowedOverlap(typ.key.length - 1) &&
-        childRVDs.forall(rvd => rvd.partitioner == childRVDs(0).partitioner)
-      )
-        childRVDs.map(_.truncateKey(typ.key.length))
-      else {
-        info("TableMultiWayZipJoin: repartitioning children")
-        val childRanges = childRVDs.flatMap(_.partitioner.coarsenedRangeBounds(typ.key.length))
-        val newPartitioner = RVDPartitioner.generate(ctx.stateManager, typ.keyType, childRanges)
-        childRVDs.map(_.repartition(ctx, newPartitioner))
-      }
-    val newPartitioner = repartitionedRVDs(0).partitioner
-
-    val rvdType = repartitionedRVDs(0).typ
-    val rowType = rvdType.rowType
-    val keyIdx = rvdType.kFieldIdx
-    val valIdx = rvdType.valueFieldIdx
-    val localRVDType = rvdType
-    val keyFields = rvdType.kType.fields.map(f => (f.name, f.typ))
-    val valueFields = rvdType.valueType.fields.map(f => (f.name, f.typ))
-    val localNewRowType = PCanonicalStruct(
-      required = true,
-      keyFields ++ Array((
-        fieldName,
-        PCanonicalArray(
-          PCanonicalStruct(required = false, valueFields: _*),
-          required = true,
-        ),
-      )): _*
-    )
-    val localDataLength = childrenSeq.length
-    val rvMerger = { (ctx: RVDContext, it: Iterator[BoxedArrayBuilder[(RegionValue, Int)]]) =>
-      val rvb = new RegionValueBuilder(sm)
-      val newRegionValue = RegionValue()
-
-      it.map { rvs =>
-        val rv = rvs(0)._1
-        rvb.set(ctx.region)
-        rvb.start(localNewRowType)
-        rvb.startStruct()
-        rvb.addFields(rowType, rv, keyIdx) // Add the key
-        rvb.startMissingArray(localDataLength) // add the values
-        var i = 0
-        while (i < rvs.length) {
-          val (rv, j) = rvs(i)
-          rvb.setArrayIndex(j)
-          rvb.setPresent()
-          rvb.startStruct()
-          rvb.addFields(rowType, rv, valIdx)
-          rvb.endStruct()
-          i += 1
-        }
-        rvb.endArrayUnchecked()
-        rvb.endStruct()
-
-        newRegionValue.set(rvb.region, rvb.end())
-        newRegionValue
-      }
-    }
-
-    val rvd = RVD(
-      typ = RVDType(localNewRowType, typ.key),
-      partitioner = newPartitioner,
-      crdd = ContextRDD.czipNPartitions(repartitionedRVDs.map(_.crdd.toCRDDRegionValue)) {
-        (ctx, its) =>
-          val orvIters = its.map(it => OrderedRVIterator(localRVDType, it, ctx, sm))
-          rvMerger(ctx, OrderedRVIterator.multiZipJoin(sm, orvIters))
-      }.toCRDDPtr,
-    )
-
-    val newGlobals = BroadcastRow(ctx, Row(childValues.map(_.globals.javaValue)), newGlobalType)
-
-    new TableValueIntermediate(TableValue(ctx, typ, newGlobals, rvd))
-  }
 }
 
 case class TableLeftJoinRightDistinct(left: TableIR, right: TableIR, root: String) extends TableIR {
@@ -2864,21 +2497,6 @@ case class TableLeftJoinRightDistinct(left: TableIR, right: TableIR, root: Strin
     : TableLeftJoinRightDistinct = {
     val IndexedSeq(newLeft: TableIR, newRight: TableIR) = newChildren
     TableLeftJoinRightDistinct(newLeft, newRight, root)
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val leftValue = left.execute(ctx, r).asTableValue(ctx)
-    val rightValue = right.execute(ctx, r).asTableValue(ctx)
-
-    val joinKey = math.min(left.typ.key.length, right.typ.key.length)
-    new TableValueIntermediate(
-      leftValue.copy(
-        typ = typ,
-        rvd = leftValue.rvd
-          .orderedLeftJoinDistinctAndInsert(rightValue.rvd.truncateKey(joinKey), root),
-      )
-    )
   }
 }
 
@@ -2933,77 +2551,6 @@ case class TableMapPartitions(
       allowedOverlap,
     )
   }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val tv = child.execute(ctx, r).asTableValue(ctx)
-    val rowPType = tv.rvd.rowPType
-    val globalPType = tv.globals.t
-
-    val (newRowPType: PStruct, makeIterator) = CompileIterator.forTableMapPartitions(
-      ctx,
-      globalPType,
-      rowPType,
-      Subst(
-        body,
-        BindingEnv(Env(
-          globalName -> In(
-            0,
-            SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(globalPType)),
-          ),
-          partitionStreamName -> In(
-            1,
-            SingleCodeEmitParamType(
-              true,
-              StreamSingleCodeType(requiresMemoryManagementPerElement = true, rowPType, true),
-            ),
-          ),
-        )),
-      ),
-    )
-
-    val globalsBc = tv.globals.broadcast(ctx.theHailClassLoader)
-
-    val fsBc = tv.ctx.fsBc
-    val itF = { (idx: Int, consumerCtx: RVDContext, partition: (RVDContext) => Iterator[Long]) =>
-      val boxedPartition = new NoBoxLongIterator {
-        var eos: Boolean = false
-        var iter: Iterator[Long] = _
-        override def init(partitionRegion: Region, elementRegion: Region): Unit =
-          iter = partition(new RVDContext(partitionRegion, elementRegion))
-
-        override def next(): Long =
-          if (!iter.hasNext) {
-            eos = true
-            0L
-          } else
-            iter.next()
-
-        override def close(): Unit = ()
-      }
-      makeIterator(
-        theHailClassLoaderForSparkWorkers,
-        fsBc.value,
-        SparkTaskContext.get(),
-        consumerCtx,
-        globalsBc.value.readRegionValue(
-          consumerCtx.partitionRegion,
-          theHailClassLoaderForSparkWorkers,
-        ),
-        boxedPartition,
-      ).map(l => l.longValue())
-    }
-
-    val rvd = tv.rvd.repartition(ctx, tv.rvd.partitioner.strictify(allowedOverlap))
-
-    new TableValueIntermediate(
-      tv.copy(
-        typ = typ,
-        rvd = rvd
-          .mapPartitionsWithContextAndIndex(RVDType(newRowPType, typ.key))(itF),
-      )
-    )
-  }
 }
 
 // Must leave key fields unchanged.
@@ -3025,383 +2572,6 @@ case class TableMapRows(child: TableIR, newRow: IR) extends TableIR {
   }
 
   override def partitionCounts: Option[IndexedSeq[Long]] = child.partitionCounts
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val tv = child.execute(ctx, r).asTableValue(ctx)
-    val fsBc = ctx.fsBc
-    val extracted = agg.Extract.apply(newRow, Requiredness(this, ctx), isScan = true)
-
-    if (extracted.aggs.isEmpty) {
-      val (Some(PTypeReferenceSingleCodeType(rTyp)), f) =
-        Compile[AsmFunction3RegionLongLongLong](
-          ctx,
-          FastSeq(
-            (
-              TableIR.globalName,
-              SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.globals.t)),
-            ),
-            (
-              TableIR.rowName,
-              SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.rvd.rowPType)),
-            ),
-          ),
-          FastSeq(classInfo[Region], LongInfo, LongInfo),
-          LongInfo,
-          Coalesce(FastSeq(
-            extracted.postAggIR,
-            Die("Internal error: TableMapRows: row expression missing", extracted.postAggIR.typ),
-          )),
-        )
-
-      val rowIterationNeedsGlobals = Mentions(extracted.postAggIR, TableIR.globalName)
-      val globalsBc =
-        if (rowIterationNeedsGlobals)
-          tv.globals.broadcast(ctx.theHailClassLoader)
-        else
-          null
-
-      val fsBc = ctx.fsBc
-      val itF = { (i: Int, ctx: RVDContext, it: Iterator[Long]) =>
-        val globalRegion = ctx.partitionRegion
-        val globals = if (rowIterationNeedsGlobals)
-          globalsBc.value.readRegionValue(globalRegion, theHailClassLoaderForSparkWorkers)
-        else
-          0
-
-        val newRow =
-          f(theHailClassLoaderForSparkWorkers, fsBc.value, SparkTaskContext.get(), globalRegion)
-        it.map(ptr => newRow(ctx.r, globals, ptr))
-      }
-
-      return new TableValueIntermediate(
-        tv.copy(
-          typ = typ,
-          rvd = tv.rvd.mapPartitionsWithIndex(RVDType(rTyp.asInstanceOf[PStruct], typ.key))(itF),
-        )
-      )
-    }
-
-    val scanInitNeedsGlobals = Mentions(extracted.init, TableIR.globalName)
-    val scanSeqNeedsGlobals = Mentions(extracted.seqPerElt, TableIR.globalName)
-    val rowIterationNeedsGlobals = Mentions(extracted.postAggIR, TableIR.globalName)
-
-    val globalsBc =
-      if (rowIterationNeedsGlobals || scanInitNeedsGlobals || scanSeqNeedsGlobals)
-        tv.globals.broadcast(ctx.theHailClassLoader)
-      else
-        null
-
-    val spec = BufferSpec.blockedUncompressed
-
-    // Order of operations:
-    // 1. init op on all aggs and serialize to byte array.
-    // 2. load in init op on each partition, seq op over partition, serialize.
-    // 3. load in partition aggregations, comb op as necessary, serialize.
-    // 4. load in partStarts, calculate newRow based on those results.
-
-    val (_, initF) = CompileWithAggregators[AsmFunction2RegionLongUnit](
-      ctx,
-      extracted.states,
-      FastSeq((
-        TableIR.globalName,
-        SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.globals.t)),
-      )),
-      FastSeq(classInfo[Region], LongInfo),
-      UnitInfo,
-      Begin(FastSeq(extracted.init)),
-    )
-
-    val serializeF = extracted.serialize(ctx, spec)
-
-    val (_, eltSeqF) = CompileWithAggregators[AsmFunction3RegionLongLongUnit](
-      ctx,
-      extracted.states,
-      FastSeq(
-        (
-          TableIR.globalName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.globals.t)),
-        ),
-        (
-          TableIR.rowName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.rvd.rowPType)),
-        ),
-      ),
-      FastSeq(classInfo[Region], LongInfo, LongInfo),
-      UnitInfo,
-      extracted.seqPerElt,
-    )
-
-    val read = extracted.deserialize(ctx, spec)
-    val write = extracted.serialize(ctx, spec)
-    val combOpFNeedsPool = extracted.combOpFSerializedFromRegionPool(ctx, spec)
-
-    val (Some(PTypeReferenceSingleCodeType(rTyp)), f) =
-      CompileWithAggregators[AsmFunction3RegionLongLongLong](
-        ctx,
-        extracted.states,
-        FastSeq(
-          (
-            TableIR.globalName,
-            SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.globals.t)),
-          ),
-          (
-            TableIR.rowName,
-            SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.rvd.rowPType)),
-          ),
-        ),
-        FastSeq(classInfo[Region], LongInfo, LongInfo),
-        LongInfo,
-        Let(
-          FastSeq(extracted.resultRef.name -> extracted.results),
-          Coalesce(FastSeq(
-            extracted.postAggIR,
-            Die("Internal error: TableMapRows: row expression missing", extracted.postAggIR.typ),
-          )),
-        ),
-      )
-    assert(rTyp.virtualType == newRow.typ)
-
-    // 1. init op on all aggs and write out to initPath
-    val initAgg = ctx.r.pool.scopedRegion { aggRegion =>
-      ctx.r.pool.scopedRegion { fRegion =>
-        val init = initF(ctx.theHailClassLoader, fsBc.value, ctx.taskContext, fRegion)
-        init.newAggState(aggRegion)
-        init(fRegion, tv.globals.value.offset)
-        serializeF(ctx.theHailClassLoader, ctx.taskContext, aggRegion, init.getAggOffset())
-      }
-    }
-
-    if (ctx.getFlag("distributed_scan_comb_op") != null && extracted.shouldTreeAggregate) {
-      val fsBc = ctx.fs.broadcast
-      val tmpBase = ctx.createTmpPath("table-map-rows-distributed-scan")
-      val d = digitsNeeded(tv.rvd.getNumPartitions)
-      val files = tv.rvd.mapPartitionsWithIndex { (i, ctx, it) =>
-        val path = tmpBase + "/" + partFile(d, i, TaskContext.get)
-        val globalRegion = ctx.freshRegion()
-        val globals = if (scanSeqNeedsGlobals)
-          globalsBc.value.readRegionValue(globalRegion, theHailClassLoaderForSparkWorkers)
-        else 0
-
-        ctx.r.pool.scopedSmallRegion { aggRegion =>
-          val tc = SparkTaskContext.get()
-          val seq = eltSeqF(theHailClassLoaderForSparkWorkers, fsBc.value, tc, globalRegion)
-
-          seq.setAggState(
-            aggRegion,
-            read(theHailClassLoaderForSparkWorkers, tc, aggRegion, initAgg),
-          )
-          it.foreach { ptr =>
-            seq(ctx.region, globals, ptr)
-            ctx.region.clear()
-          }
-          using(new DataOutputStream(fsBc.value.create(path))) { os =>
-            val bytes = write(theHailClassLoaderForSparkWorkers, tc, aggRegion, seq.getAggOffset())
-            os.writeInt(bytes.length)
-            os.write(bytes)
-          }
-          Iterator.single(path)
-        }
-      }.collect()
-
-      val fileStack = new BoxedArrayBuilder[Array[String]]()
-      var filesToMerge: Array[String] = files
-      while (filesToMerge.length > 1) {
-        val nToMerge = filesToMerge.length / 2
-        log.info(s"Running distributed combine stage with $nToMerge tasks")
-        fileStack += filesToMerge
-
-        filesToMerge =
-          ContextRDD.weaken(SparkBackend.sparkContext("TableMapRows.execute").parallelize(
-            0 until nToMerge,
-            nToMerge,
-          ))
-            .cmapPartitions { (ctx, it) =>
-              val i = it.next()
-              assert(it.isEmpty)
-              val path = tmpBase + "/" + partFile(d, i, TaskContext.get)
-              val file1 = filesToMerge(i * 2)
-              val file2 = filesToMerge(i * 2 + 1)
-
-              def readToBytes(is: DataInputStream): Array[Byte] = {
-                val len = is.readInt()
-                val b = new Array[Byte](len)
-                is.readFully(b)
-                b
-              }
-
-              val b1 = using(new DataInputStream(fsBc.value.open(file1)))(readToBytes)
-              val b2 = using(new DataInputStream(fsBc.value.open(file2)))(readToBytes)
-              using(new DataOutputStream(fsBc.value.create(path))) { os =>
-                val bytes = combOpFNeedsPool(() =>
-                  (ctx.r.pool, theHailClassLoaderForSparkWorkers, SparkTaskContext.get())
-                )(b1, b2)
-                os.writeInt(bytes.length)
-                os.write(bytes)
-              }
-              Iterator.single(path)
-            }.collect()
-      }
-      fileStack += filesToMerge
-
-      val itF = { (i: Int, ctx: RVDContext, it: Iterator[Long]) =>
-        val globalRegion = ctx.freshRegion()
-        val globals = if (rowIterationNeedsGlobals || scanSeqNeedsGlobals)
-          globalsBc.value.readRegionValue(globalRegion, theHailClassLoaderForSparkWorkers)
-        else
-          0
-        val partitionAggs = {
-          var j = 0
-          var x = i
-          val ab = new BoxedArrayBuilder[String]
-          while (j < fileStack.length) {
-            assert(x <= fileStack(j).length)
-            if (x % 2 != 0) {
-              x -= 1
-              ab += fileStack(j)(x)
-            }
-            assert(x % 2 == 0)
-            x = x / 2
-            j += 1
-          }
-          assert(x == 0)
-          var b = initAgg
-          ab.result().reverseIterator.foreach { path =>
-            def readToBytes(is: DataInputStream): Array[Byte] = {
-              val len = is.readInt()
-              val b = new Array[Byte](len)
-              is.readFully(b)
-              b
-            }
-
-            b = combOpFNeedsPool(() =>
-              (ctx.r.pool, theHailClassLoaderForSparkWorkers, SparkTaskContext.get())
-            )(b, using(new DataInputStream(fsBc.value.open(path)))(readToBytes))
-          }
-          b
-        }
-
-        val aggRegion = ctx.freshRegion()
-        val hcl = theHailClassLoaderForSparkWorkers
-        val tc = SparkTaskContext.get()
-        val newRow = f(hcl, fsBc.value, tc, globalRegion)
-        val seq = eltSeqF(hcl, fsBc.value, tc, globalRegion)
-        var aggOff = read(hcl, tc, aggRegion, partitionAggs)
-
-        val res = it.map { ptr =>
-          newRow.setAggState(aggRegion, aggOff)
-          val newPtr = newRow(ctx.region, globals, ptr)
-          aggOff = newRow.getAggOffset()
-          seq.setAggState(aggRegion, aggOff)
-          seq(ctx.region, globals, ptr)
-          aggOff = seq.getAggOffset()
-          newPtr
-        }
-        res
-      }
-      return new TableValueIntermediate(
-        tv.copy(
-          typ = typ,
-          rvd = tv.rvd.mapPartitionsWithIndex(RVDType(rTyp.asInstanceOf[PStruct], typ.key))(itF),
-        )
-      )
-    }
-
-    // 2. load in init op on each partition, seq op over partition, write out.
-    val scanPartitionAggs = SpillingCollectIterator(
-      ctx.localTmpdir,
-      ctx.fs,
-      tv.rvd.mapPartitionsWithIndex { (i, ctx, it) =>
-        val globalRegion = ctx.partitionRegion
-        val globals = if (scanSeqNeedsGlobals)
-          globalsBc.value.readRegionValue(globalRegion, theHailClassLoaderForSparkWorkers)
-        else 0
-
-        SparkTaskContext.get().getRegionPool().scopedSmallRegion { aggRegion =>
-          val hcl = theHailClassLoaderForSparkWorkers
-          val tc = SparkTaskContext.get()
-          val seq = eltSeqF(hcl, fsBc.value, tc, globalRegion)
-
-          seq.setAggState(aggRegion, read(hcl, tc, aggRegion, initAgg))
-          it.foreach { ptr =>
-            seq(ctx.region, globals, ptr)
-            ctx.region.clear()
-          }
-          Iterator.single(write(hcl, tc, aggRegion, seq.getAggOffset()))
-        }
-      },
-      ctx.getFlag("max_leader_scans").toInt,
-    )
-
-    // 3. load in partition aggregations, comb op as necessary, write back out.
-    val partAggs = scanPartitionAggs.scanLeft(initAgg)(combOpFNeedsPool(() =>
-      (ctx.r.pool, ctx.theHailClassLoader, ctx.taskContext)
-    ))
-    val scanAggCount = tv.rvd.getNumPartitions
-    val partitionIndices = new Array[Long](scanAggCount)
-    val scanAggsPerPartitionFile = ctx.createTmpPath("table-map-rows-scan-aggs-part")
-    using(ctx.fs.createNoCompression(scanAggsPerPartitionFile)) { os =>
-      partAggs.zipWithIndex.foreach { case (x, i) =>
-        if (i < scanAggCount) {
-          log.info(s"TableMapRows scan: serializing combined agg $i")
-          partitionIndices(i) = os.getPosition
-          os.writeInt(x.length)
-          os.write(x, 0, x.length)
-        }
-      }
-    }
-
-    // 4. load in partStarts, calculate newRow based on those results.
-    val itF = { (i: Int, ctx: RVDContext, filePosition: Long, it: Iterator[Long]) =>
-      val globalRegion = ctx.partitionRegion
-      val globals = if (rowIterationNeedsGlobals || scanSeqNeedsGlobals)
-        globalsBc.value.readRegionValue(globalRegion, theHailClassLoaderForSparkWorkers)
-      else
-        0
-      val partitionAggs = using(fsBc.value.openNoCompression(scanAggsPerPartitionFile)) { is =>
-        is.seek(filePosition)
-        val aggSize = is.readInt()
-        val partAggs = new Array[Byte](aggSize)
-        var nread = is.read(partAggs, 0, aggSize)
-        var r = nread
-        while (r > 0 && nread < aggSize) {
-          r = is.read(partAggs, nread, aggSize - nread)
-          if (r > 0) nread += r
-        }
-        if (nread != aggSize) {
-          fatal(s"aggs read wrong number of bytes: $nread vs $aggSize")
-        }
-        partAggs
-      }
-
-      val aggRegion = ctx.freshRegion()
-      val hcl = theHailClassLoaderForSparkWorkers
-      val tc = SparkTaskContext.get()
-      val newRow = f(hcl, fsBc.value, tc, globalRegion)
-      val seq = eltSeqF(hcl, fsBc.value, tc, globalRegion)
-      var aggOff = read(hcl, tc, aggRegion, partitionAggs)
-
-      var idx = 0
-      it.map { ptr =>
-        newRow.setAggState(aggRegion, aggOff)
-        val off = newRow(ctx.region, globals, ptr)
-        seq.setAggState(aggRegion, newRow.getAggOffset())
-        idx += 1
-        seq(ctx.region, globals, ptr)
-        aggOff = seq.getAggOffset()
-        off
-      }
-    }
-    new TableValueIntermediate(
-      tv.copy(
-        typ = typ,
-        rvd = tv.rvd.mapPartitionsWithIndexAndValue(
-          RVDType(rTyp.asInstanceOf[PStruct], typ.key),
-          partitionIndices,
-        )(itF),
-      )
-    )
-  }
 }
 
 case class TableMapGlobals(child: TableIR, newGlobals: IR) extends TableIR {
@@ -3418,32 +2588,6 @@ case class TableMapGlobals(child: TableIR, newGlobals: IR) extends TableIR {
   }
 
   override def partitionCounts: Option[IndexedSeq[Long]] = child.partitionCounts
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val tv = child.execute(ctx, r).asTableValue(ctx)
-
-    val (Some(PTypeReferenceSingleCodeType(resultPType: PStruct)), f) =
-      Compile[AsmFunction2RegionLongLong](
-        ctx,
-        FastSeq((
-          TableIR.globalName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(tv.globals.t)),
-        )),
-        FastSeq(classInfo[Region], LongInfo),
-        LongInfo,
-        Coalesce(FastSeq(
-          newGlobals,
-          Die("Internal error: TableMapGlobals: globals missing", newGlobals.typ),
-        )),
-      )
-
-    val resultOff =
-      f(ctx.theHailClassLoader, ctx.fs, ctx.taskContext, ctx.r)(ctx.r, tv.globals.value.offset)
-    TableValueIntermediate(
-      tv.copy(typ = typ, globals = BroadcastRow(ctx, RegionValue(ctx.r, resultOff), resultPType))
-    )
-  }
 }
 
 case class TableExplode(child: TableIR, path: IndexedSeq[String]) extends TableIR {
@@ -3458,106 +2602,16 @@ case class TableExplode(child: TableIR, path: IndexedSeq[String]) extends TableI
 
   private def childRowType = child.typ.rowType
 
-  private[this] lazy val idx = Ref(freshName(), TInt32)
-
-  private[this] lazy val newRow: InsertFields = {
-    val refs = path.init.scanLeft(Ref(TableIR.rowName, childRowType))((struct, name) =>
-      Ref(freshName(), tcoerce[TStruct](struct.typ).field(name).typ)
+  lazy val typ: TableType = child.typ.copy(rowType =
+    childRowType.structInsert(
+      tcoerce[TContainer](childRowType.queryTyped(path)._1).elementType,
+      path,
     )
-
-    path.zip(refs).zipWithIndex.foldRight[IR](idx) {
-      case (((field, ref), i), arg) =>
-        InsertFields(
-          ref,
-          FastSeq(field ->
-            (if (i == refs.length - 1)
-               ArrayRef(CastToArray(GetField(ref, field)), arg)
-             else
-               Let(FastSeq(refs(i + 1).name -> GetField(ref, field)), arg))),
-        )
-    }.asInstanceOf[InsertFields]
-  }
-
-  lazy val typ: TableType = child.typ.copy(rowType = newRow.typ)
+  )
 
   override protected def copyWithNewChildren(newChildren: IndexedSeq[BaseIR]): TableExplode = {
     assert(newChildren.length == 1)
     TableExplode(newChildren(0).asInstanceOf[TableIR], path)
-  }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val prev = child.execute(ctx, r).asTableValue(ctx)
-
-    val length: IR =
-      Coalesce(FastSeq(
-        ArrayLen(CastToArray(
-          path.foldLeft[IR](Ref(TableIR.rowName, childRowType))((struct, field) =>
-            GetField(struct, field)
-          )
-        )),
-        0,
-      ))
-
-    val (_, l) = Compile[AsmFunction2RegionLongInt](
-      ctx,
-      FastSeq((
-        TableIR.rowName,
-        SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.rvd.rowPType)),
-      )),
-      FastSeq(classInfo[Region], LongInfo),
-      IntInfo,
-      length,
-    )
-    val (Some(PTypeReferenceSingleCodeType(newRowType: PStruct)), f) =
-      Compile[AsmFunction3RegionLongIntLong](
-        ctx,
-        FastSeq(
-          (
-            TableIR.rowName,
-            SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.rvd.rowPType)),
-          ),
-          (idx.name, SingleCodeEmitParamType(true, Int32SingleCodeType)),
-        ),
-        FastSeq(classInfo[Region], LongInfo, IntInfo),
-        LongInfo,
-        newRow,
-      )
-    assert(newRowType.virtualType == typ.rowType)
-
-    val rvdType: RVDType = RVDType(
-      newRowType,
-      prev.rvd.typ.key.takeWhile(_ != path.head),
-    )
-    val fsBc = ctx.fsBc
-    TableValueIntermediate(
-      TableValue(
-        ctx,
-        typ,
-        prev.globals,
-        prev.rvd.boundary.mapPartitionsWithIndex(rvdType) { (i, ctx, it) =>
-          val globalRegion = ctx.partitionRegion
-          val lenF =
-            l(theHailClassLoaderForSparkWorkers, fsBc.value, SparkTaskContext.get(), globalRegion)
-          val rowF =
-            f(theHailClassLoaderForSparkWorkers, fsBc.value, SparkTaskContext.get(), globalRegion)
-          it.flatMap { ptr =>
-            val len = lenF(ctx.region, ptr)
-            new Iterator[Long] {
-              private[this] var i = 0
-
-              def hasNext: Boolean = i < len
-
-              def next(): Long = {
-                val ret = rowF(ctx.region, ptr, i)
-                i += 1
-                ret
-              }
-            }
-          }
-        },
-      )
-    )
   }
 }
 
@@ -3581,16 +2635,6 @@ case class TableUnion(childrenSeq: IndexedSeq[TableIR]) extends TableIR {
     TableUnion(newChildren.map(_.asInstanceOf[TableIR]))
 
   def typ: TableType = childrenSeq(0).typ
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val tvs = childrenSeq.map(_.execute(ctx, r).asTableValue(ctx))
-    TableValueIntermediate(
-      tvs(0).copy(
-        rvd = RVD.union(RVD.unify(ctx, tvs.map(_.rvd)), tvs(0).typ.key.length, ctx)
-      )
-    )
-  }
 }
 
 case class MatrixRowsTable(child: MatrixIR) extends TableIR {
@@ -3646,14 +2690,6 @@ case class TableDistinct(child: TableIR) extends TableIR {
   }
 
   def typ: TableType = child.typ
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val prev = child.execute(ctx, r).asTableValue(ctx)
-    new TableValueIntermediate(prev.copy(rvd =
-      prev.rvd.truncateKey(prev.typ.key).distinctByKey(ctx)
-    ))
-  }
 }
 
 case class TableKeyByAndAggregate(
@@ -3687,185 +2723,6 @@ case class TableKeyByAndAggregate(
     globalType = child.typ.globalType,
     key = keyType.fieldNames,
   )
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val prev = child.execute(ctx, r).asTableValue(ctx)
-    val fsBc = ctx.fsBc
-    val sm = ctx.stateManager
-
-    val localKeyType = keyType
-    val (Some(PTypeReferenceSingleCodeType(localKeyPType: PStruct)), makeKeyF) =
-      Compile[AsmFunction3RegionLongLongLong](
-        ctx,
-        FastSeq(
-          (
-            TableIR.rowName,
-            SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.rvd.rowPType)),
-          ),
-          (
-            TableIR.globalName,
-            SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.globals.t)),
-          ),
-        ),
-        FastSeq(classInfo[Region], LongInfo, LongInfo),
-        LongInfo,
-        Coalesce(FastSeq(
-          newKey,
-          Die("Internal error: TableKeyByAndAggregate: newKey missing", newKey.typ),
-        )),
-      )
-
-    val globalsBc = prev.globals.broadcast(ctx.theHailClassLoader)
-
-    val spec = BufferSpec.blockedUncompressed
-
-    val extracted = agg.Extract(expr, Requiredness(this, ctx))
-
-    val (_, makeInit) = CompileWithAggregators[AsmFunction2RegionLongUnit](
-      ctx,
-      extracted.states,
-      FastSeq((
-        TableIR.globalName,
-        SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.globals.t)),
-      )),
-      FastSeq(classInfo[Region], LongInfo),
-      UnitInfo,
-      extracted.init,
-    )
-
-    val (_, makeSeq) = CompileWithAggregators[AsmFunction3RegionLongLongUnit](
-      ctx,
-      extracted.states,
-      FastSeq(
-        (
-          TableIR.globalName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.globals.t)),
-        ),
-        (
-          TableIR.rowName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.rvd.rowPType)),
-        ),
-      ),
-      FastSeq(classInfo[Region], LongInfo, LongInfo),
-      UnitInfo,
-      extracted.seqPerElt,
-    )
-
-    val (Some(PTypeReferenceSingleCodeType(rTyp: PStruct)), makeAnnotate) =
-      CompileWithAggregators[AsmFunction2RegionLongLong](
-        ctx,
-        extracted.states,
-        FastSeq((
-          TableIR.globalName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.globals.t)),
-        )),
-        FastSeq(classInfo[Region], LongInfo),
-        LongInfo,
-        Let(FastSeq(extracted.resultRef.name -> extracted.results), extracted.postAggIR),
-      )
-    assert(rTyp.virtualType == typ.valueType, s"$rTyp, ${typ.valueType}")
-
-    val serialize = extracted.serialize(ctx, spec)
-    val deserialize = extracted.deserialize(ctx, spec)
-    val combOp = extracted.combOpFSerializedWorkersOnly(ctx, spec)
-
-    val hcl = theHailClassLoaderForSparkWorkers
-    val tc = ctx.taskContext
-    val initF = makeInit(hcl, fsBc.value, tc, ctx.r)
-    val globalsOffset = prev.globals.value.offset
-    val initAggs = ctx.r.pool.scopedRegion { aggRegion =>
-      initF.newAggState(aggRegion)
-      initF(ctx.r, globalsOffset)
-      serialize(hcl, tc, aggRegion, initF.getAggOffset())
-    }
-
-    val newRowType = PCanonicalStruct(
-      required = true,
-      localKeyPType.fields.map(f => (f.name, PType.canonical(f.typ))) ++ rTyp.fields.map(f =>
-        (f.name, f.typ)
-      ): _*
-    )
-
-    val localBufferSize = bufferSize
-    val rdd = prev.rvd
-      .boundary
-      .mapPartitionsWithIndex { (i, ctx, it) =>
-        val partRegion = ctx.partitionRegion
-        val hcl = theHailClassLoaderForSparkWorkers
-        val tc = SparkTaskContext.get()
-        val globals = globalsBc.value.readRegionValue(partRegion, hcl)
-        val makeKey = {
-          val f = makeKeyF(hcl, fsBc.value, tc, partRegion)
-          ptr: Long => {
-            val keyOff = f(ctx.region, ptr, globals)
-            SafeRow.read(localKeyPType, keyOff).asInstanceOf[Row]
-          }
-        }
-        val makeAgg = { () =>
-          val aggRegion = ctx.freshRegion()
-          RegionValue(aggRegion, deserialize(hcl, tc, aggRegion, initAggs))
-        }
-
-        val seqOp = {
-          val f = makeSeq(hcl, fsBc.value, SparkTaskContext.get(), partRegion)
-          (ptr: Long, agg: RegionValue) => {
-            f.setAggState(agg.region, agg.offset)
-            f(ctx.region, globals, ptr)
-            agg.setOffset(f.getAggOffset())
-            ctx.region.clear()
-          }
-        }
-        val serializeAndCleanupAggs = { rv: RegionValue =>
-          val a = serialize(hcl, tc, rv.region, rv.offset)
-          rv.region.close()
-          a
-        }
-
-        new BufferedAggregatorIterator[Long, RegionValue, Array[Byte], Row](
-          it,
-          makeAgg,
-          makeKey,
-          seqOp,
-          serializeAndCleanupAggs,
-          localBufferSize)
-      }.aggregateByKey(initAggs, nPartitions.getOrElse(prev.rvd.getNumPartitions))(combOp, combOp)
-
-    val crdd = ContextRDD.weaken(rdd).cmapPartitionsWithIndex({ (i, ctx, it) =>
-      val region = ctx.region
-
-      val rvb = new RegionValueBuilder(sm)
-      val partRegion = ctx.partitionRegion
-      val hcl = theHailClassLoaderForSparkWorkers
-      val tc = SparkTaskContext.get()
-      val globals = globalsBc.value.readRegionValue(partRegion, hcl)
-      val annotate = makeAnnotate(hcl, fsBc.value, tc, partRegion)
-
-      it.map { case (key, aggs) =>
-        rvb.set(region)
-        rvb.start(newRowType)
-        rvb.startStruct()
-        var i = 0
-        while (i < localKeyType.size) {
-          rvb.addAnnotation(localKeyType.types(i), key.get(i))
-          i += 1
-        }
-
-        val aggOff = deserialize(hcl, tc, region, aggs)
-        annotate.setAggState(region, aggOff)
-        rvb.addAllFields(rTyp, region, annotate(region, globals))
-        rvb.endStruct()
-        rvb.end()
-      }
-    })
-
-    new TableValueIntermediate(
-      prev.copy(
-        typ = typ,
-        rvd = RVD.coerce(ctx, RVDType(newRowType, keyType.fieldNames), crdd),
-      )
-    )
-  }
 }
 
 // follows key_by non-empty key
@@ -3886,142 +2743,6 @@ case class TableAggregateByKey(child: TableIR, expr: IR) extends TableIR {
 
   lazy val typ: TableType =
     child.typ.copy(rowType = child.typ.keyType ++ tcoerce[TStruct](expr.typ))
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val prev = child.execute(ctx, r).asTableValue(ctx)
-    val prevRVD = prev.rvd.truncateKey(child.typ.key)
-    val fsBc = ctx.fsBc
-    val sm = ctx.stateManager
-
-    val extracted = agg.Extract(expr, Requiredness(this, ctx))
-
-    val (_, makeInit) = CompileWithAggregators[AsmFunction2RegionLongUnit](
-      ctx,
-      extracted.states,
-      FastSeq((
-        TableIR.globalName,
-        SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.globals.t)),
-      )),
-      FastSeq(classInfo[Region], LongInfo),
-      UnitInfo,
-      extracted.init,
-    )
-
-    val (_, makeSeq) = CompileWithAggregators[AsmFunction3RegionLongLongUnit](
-      ctx,
-      extracted.states,
-      FastSeq(
-        (
-          TableIR.globalName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.globals.t)),
-        ),
-        (
-          TableIR.rowName,
-          SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prevRVD.rowPType)),
-        ),
-      ),
-      FastSeq(classInfo[Region], LongInfo, LongInfo),
-      UnitInfo,
-      extracted.seqPerElt,
-    )
-
-    val valueIR = Let(FastSeq(extracted.resultRef.name -> extracted.results), extracted.postAggIR)
-    val keyType = prevRVD.typ.kType
-
-    val key = Ref(freshName(), keyType.virtualType)
-    val value = Ref(freshName(), valueIR.typ)
-    val (Some(PTypeReferenceSingleCodeType(rowType: PStruct)), makeRow) =
-      CompileWithAggregators[AsmFunction3RegionLongLongLong](
-        ctx,
-        extracted.states,
-        FastSeq(
-          (
-            TableIR.globalName,
-            SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(prev.globals.t)),
-          ),
-          (key.name, SingleCodeEmitParamType(true, PTypeReferenceSingleCodeType(keyType))),
-        ),
-        FastSeq(classInfo[Region], LongInfo, LongInfo),
-        LongInfo,
-        Let(
-          FastSeq(value.name -> valueIR),
-          InsertFields(key, typ.valueType.fieldNames.map(n => n -> GetField(value, n))),
-        ),
-      )
-
-    assert(rowType.virtualType == typ.rowType, s"$rowType, ${typ.rowType}")
-
-    val localChildRowType = prevRVD.rowPType
-    val keyIndices = prevRVD.typ.kFieldIdx
-    val keyOrd = prevRVD.typ.kRowOrd(ctx.stateManager)
-    val globalsBc = prev.globals.broadcast(ctx.theHailClassLoader)
-
-    val newRVDType = prevRVD.typ.copy(rowType = rowType)
-
-    val newRVD = prevRVD
-      .repartition(ctx, prevRVD.partitioner.strictify())
-      .boundary
-      .mapPartitionsWithIndex(newRVDType) { (i, ctx, it) =>
-        val partRegion = ctx.partitionRegion
-        val globalsOff =
-          globalsBc.value.readRegionValue(partRegion, theHailClassLoaderForSparkWorkers)
-
-        val initialize = makeInit(
-          theHailClassLoaderForSparkWorkers,
-          fsBc.value,
-          SparkTaskContext.get(),
-          partRegion,
-        )
-        val sequence =
-          makeSeq(theHailClassLoaderForSparkWorkers, fsBc.value, SparkTaskContext.get(), partRegion)
-        val newRowF =
-          makeRow(theHailClassLoaderForSparkWorkers, fsBc.value, SparkTaskContext.get(), partRegion)
-
-        val aggRegion = ctx.freshRegion()
-
-        new Iterator[Long] {
-          var isEnd = false
-          var current: Long = 0
-          val rowKey: WritableRegionValue = WritableRegionValue(sm, keyType, ctx.freshRegion())
-          val consumerRegion: Region = ctx.region
-
-          def hasNext: Boolean = {
-            if (isEnd || (current == 0 && !it.hasNext)) {
-              isEnd = true
-              return false
-            }
-            if (current == 0)
-              current = it.next()
-            true
-          }
-
-          def next(): Long = {
-            if (!hasNext)
-              throw new java.util.NoSuchElementException()
-
-            rowKey.setSelect(localChildRowType, keyIndices, current, true)
-
-            aggRegion.clear()
-            initialize.newAggState(aggRegion)
-            initialize(ctx.r, globalsOff)
-            sequence.setAggState(aggRegion, initialize.getAggOffset())
-
-            do {
-              sequence(ctx.r, globalsOff, current)
-              current = 0
-            } while (hasNext && keyOrd.equiv(rowKey.value.offset, current))
-            newRowF.setAggState(aggRegion, sequence.getAggOffset())
-
-            newRowF(consumerRegion, globalsOff, rowKey.offset)
-          }
-        }
-      }
-
-    new TableValueIntermediate(
-      prev.copy(rvd = newRVD, typ = typ)
-    )
-  }
 }
 
 object TableOrderBy {
@@ -4033,7 +2754,6 @@ object TableOrderBy {
 }
 
 case class TableOrderBy(child: TableIR, sortFields: IndexedSeq[SortField]) extends TableIR {
-
   lazy val definitelyDoesNotShuffle: Boolean =
     TableOrderBy.isAlreadyOrdered(sortFields, child.typ.key)
   // TableOrderBy expects an unkeyed child, so that we can better optimize by
@@ -4049,37 +2769,6 @@ case class TableOrderBy(child: TableIR, sortFields: IndexedSeq[SortField]) exten
   }
 
   lazy val typ: TableType = child.typ.copy(key = FastSeq())
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val prev = child.execute(ctx, r).asTableValue(ctx)
-
-    val physicalKey = prev.rvd.typ.key
-    if (TableOrderBy.isAlreadyOrdered(sortFields, physicalKey))
-      return new TableValueIntermediate(prev.copy(typ = typ))
-
-    val rowType = child.typ.rowType
-    val sortColIndexOrd = sortFields.map { case SortField(n, so) =>
-      val i = rowType.fieldIdx(n)
-      val f = rowType.fields(i)
-      val fo = f.typ.ordering(ctx.stateManager)
-      if (so == Ascending) fo else fo.reverse
-    }.toArray
-
-    val ord: Ordering[Annotation] = ExtendedOrdering.rowOrdering(sortColIndexOrd).toOrdering
-
-    val act = implicitly[ClassTag[Annotation]]
-
-    val codec = TypedCodecSpec(ctx, prev.rvd.rowPType, BufferSpec.wireSpec)
-    val rdd = prev.rvd.keyedEncodedRDD(ctx, codec, sortFields.map(_.field)).sortBy(_._1)(ord, act)
-    val (rowPType: PStruct, orderedCRDD) = codec.decodeRDD(ctx, rowType, rdd.map(_._2))
-    new TableValueIntermediate(TableValue(
-      ctx,
-      typ,
-      prev.globals,
-      RVD.unkeyed(rowPType, orderedCRDD),
-    ))
-  }
 }
 
 /** Create a Table from a MatrixTable, storing the column values in a global field 'colsFieldName',
@@ -4130,12 +2819,6 @@ case class TableRename(child: TableIR, rowMap: Map[String, String], globalMap: M
     val IndexedSeq(newChild: TableIR) = newChildren
     TableRename(newChild, rowMap, globalMap)
   }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate =
-    TableValueIntermediate(
-      child.execute(ctx, r).asTableValue(ctx).rename(globalMap, rowMap)
-    )
 }
 
 case class TableFilterIntervals(child: TableIR, intervals: IndexedSeq[Interval], keep: Boolean)
@@ -4150,20 +2833,6 @@ case class TableFilterIntervals(child: TableIR, intervals: IndexedSeq[Interval],
   }
 
   override def typ: TableType = child.typ
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val tv = child.execute(ctx, r).asTableValue(ctx)
-    val partitioner = RVDPartitioner.union(
-      ctx.stateManager,
-      tv.typ.keyType,
-      intervals,
-      tv.typ.keyType.size - 1,
-    )
-    new TableValueIntermediate(
-      TableValue(ctx, tv.typ, tv.globals, tv.rvd.filterIntervals(partitioner, keep))
-    )
-  }
 }
 
 case class MatrixToTableApply(child: MatrixIR, function: MatrixToTableFunction) extends TableIR {
@@ -4198,10 +2867,6 @@ case class TableToTableApply(child: TableIR, function: TableToTableFunction) ext
 
   lazy val rowCountUpperBound: Option[Long] =
     if (function.preservesPartitionCounts) child.rowCountUpperBound else None
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate =
-    new TableValueIntermediate(function.execute(ctx, child.execute(ctx, r).asTableValue(ctx)))
 }
 
 case class BlockMatrixToTableApply(
@@ -4222,13 +2887,6 @@ case class BlockMatrixToTableApply(
     )
 
   override lazy val typ: TableType = function.typ(bm.typ, aux.typ)
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate = {
-    val b = bm.execute(ctx)
-    val a = CompileAndEvaluate[Any](ctx, aux, optimize = false)
-    new TableValueIntermediate(function.execute(ctx, b, a))
-  }
 }
 
 case class BlockMatrixToTable(child: BlockMatrixIR) extends TableIR {
@@ -4245,10 +2903,6 @@ case class BlockMatrixToTable(child: BlockMatrixIR) extends TableIR {
     val rvType = TStruct("i" -> TInt64, "j" -> TInt64, "entry" -> TFloat64)
     TableType(rvType, Array[String](), TStruct.empty)
   }
-
-  override protected[ir] def execute(ctx: ExecuteContext, r: LoweringAnalyses)
-    : TableExecuteIntermediate =
-    TableValueIntermediate(child.execute(ctx).entriesTable(ctx))
 }
 
 case class RelationalLetTable(name: Name, value: IR, body: TableIR) extends TableIR {
